Removed length from method

+++ nuttx/configs/stm32f777zit6-meadow/src/espcp/espcp_file_system.h
@@ -102,8 +102,8 @@ typedef struct espcp_file_system_info_s espcp_file_system_info_t;
  ****************************************************************************/
 
 int espcp_file_system_format(void);
-uint8_t *espcp_file_system_read_file(char *name, uint32_t *length);
-int espcp_file_system_write_file(char *name, uint8_t *buffer, uint16_t length);
+uint8_t *espcp_file_system_read_file(char *name, int16_t *length);
+int espcp_file_system_write_file(char *name, uint8_t *buffer, int16_t length);
 int espcp_file_system_delete_file(char *name);
 espcp_file_system_info_t *espcp_file_system_list_files(void);
 void espcp_file_system_info_dispose(espcp_file_system_info_t *);
